Move log service into chroma-log-service crate.
