+++ Cargo.toml
@@ -1,7 +1,34 @@
 [workspace]
 resolver = "2"
 
-members = ["rust/benchmark", "rust/blockstore", "rust/cache", "rust/chroma", "rust/config", "rust/distance", "rust/error", "rust/frontend", "rust/garbage_collector", "rust/index", "rust/load", "rust/log", "rust/memberlist", "rust/storage", "rust/system", "rust/sysdb", "rust/types", "rust/worker", "rust/segment", "rust/python_bindings", "rust/mdac", "rust/tracing", "rust/sqlite", "rust/cli", "rust/wal3"]
+members = [
+    "rust/benchmark",
+    "rust/blockstore",
+    "rust/cache",
+    "rust/chroma",
+    "rust/config",
+    "rust/distance",
+    "rust/error",
+    "rust/frontend",
+    "rust/garbage_collector",
+    "rust/index",
+    "rust/load",
+    "rust/log",
+    "rust/log-service",
+    "rust/memberlist",
+    "rust/storage",
+    "rust/system",
+    "rust/sysdb",
+    "rust/types",
+    "rust/worker",
+    "rust/segment",
+    "rust/python_bindings",
+    "rust/mdac",
+    "rust/tracing",
+    "rust/sqlite",
+    "rust/cli",
+    "rust/wal3",
+]
 
 [workspace.dependencies]
 arrow = "52.2.0"
