+++ rust/log-service/Cargo.toml
@@ -1,13 +1,19 @@
 [package]
-name = "chroma-log"
+name = "chroma-log-service"
 version = "0.1.0"
 edition = "2021"
+
+[[bin]]
+name = "log_service"
+path = "src/bin/log.rs"
 
 [dependencies]
 arrow = { workspace = true }
 async-trait = { workspace = true }
 bytes = { workspace = true }
+parquet = { workspace = true }
 prost = { workspace = true }
+figment = { workspace = true }
 # Used by tracing
 opentelemetry = { workspace = true }
 rand = { workspace = true }
@@ -26,17 +32,17 @@ serde_json = { workspace = true }
 bytemuck = { workspace = true }
 futures = { workspace = true }
 
-chroma-tracing = { workspace = true, features = ["grpc"] }
 chroma-config = { workspace = true }
 chroma-error = { workspace = true, features = ["sqlx"] }
+chroma-log = { workspace = true }
 chroma-segment = { workspace = true }
-chroma-storage = { workspace = true }
-chroma-system = { workspace = true }
-chroma-sysdb = { workspace = true }
-chroma-types = { workspace = true }
 chroma-sqlite = { workspace = true }
+chroma-storage = { workspace = true }
+chroma-sysdb = { workspace = true }
+chroma-system = { workspace = true }
+chroma-tracing = { workspace = true, features = ["grpc"] }
+chroma-types = { workspace = true }
 wal3 = { workspace = true }
-figment.workspace = true
 
 [dev-dependencies]
 tempfile = { workspace = true }
