+++ rust/log-service/src/bin/log.rs
@@ -1,0 +1,4 @@
+#[tokio::main]
+async fn main() {
+    chroma_log_service::log_entrypoint().await;
+}
