--- rust/log/src/bin/log.rs
@@ -1,5 +1,0 @@
-#[tokio::main]
-async fn main() {
-    #[cfg(feature = "server")]
-    chroma_log::log_entrypoint().await;
-}
